Test fixes.
